Add Policies to Todo and TodoTask

+++ app/Http/Controllers/TodoController.php
@@ -24,6 +24,7 @@ class TodoController extends Controller
 
     public function show(Todo $todo)
     {
+        $this->authorize('view', $todo);
         $todo->load('tasks');
         return new TodoResource($todo);
     }
@@ -38,6 +39,7 @@ class TodoController extends Controller
 
     public function update(Todo $todo, TodoUpdateRequest $request)
     {
+        $this->authorize('update', $todo);
         $input = $request->validated();
         $todo->fill($input)->save();
 
@@ -46,11 +48,13 @@ class TodoController extends Controller
 
     public function destroy(Todo $todo)
     {
+        $this->authorize('destroy', $todo);
         $todo->delete();
     }
 
     public function addTask(Todo $todo, TodoTaskStoreRequest $request) 
     {
+        $this->authorize('addTask', $todo);
         $input = $request->validated();
         $todoTask = $todo->tasks()->create($input);
