+++ app/Http/Controllers/TodoTaskController.php
@@ -16,6 +16,7 @@ class TodoTaskController extends Controller
 
     public function update(TodoTask $todoTask, TodoTaskUpdateRequest $request)
     {
+        $this->authorize('update', $todoTask);
         $input = $request->validated();
         $todoTask->fill($input);
         $todoTask->save();
@@ -25,6 +26,7 @@ class TodoTaskController extends Controller
 
     public function destroy(TodoTask $todoTask)
     {
+        $this->authorize('delete', $todoTask);
         $todoTask->delete();
     }
 }
